Allow for collecting credentials from environment variable MYSQL_DSN
It may be convenient to connect to different servers, but setting
the credentials via command line parameters may not be convenient
or secure so we use an environment variable which can be set and
which ps-top will use.

This allows you to do something like:

$ MYSQL_DSN='test_user:test_pass@tcp(127.0.0.1:3306)/performance_schema' ps-top

or

$ export MYSQL_DSN='test_user:test_pass@tcp(127.0.0.1:3306)/performance_schema'
$ ps-top

Note: for now it's necessary to specify the database and port even if they
are default values. I may consider fixing this later.